pyproject.toml
[project]
name = "dlt-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dlt[postgres]==1.18.2",
    "notebook==7.2.2",
    "psycopg[binary]>=3.1.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0",
    "sqlmodel==0.0.27",
]
